refactor(estimation): dedup IMPMAP/SAEM EM scaffolding + cut hot-loop allocs [closes #273]
Maintainability + efficiency follow-up from the IMPMAP code review (#272/#270).
Correctness findings were already fixed in #272; this is the dedicated refactor.

Reuse — extract the EM-family leaf helpers that IMPMAP had copied verbatim from
SAEM (untested copies that could silently drift) into a shared `pub(crate)`
`estimation::em_common`:
  - `floor_omega_diagonal` (Ω diagonal PD floor, FIX-aware)
  - `get_mu_ref_pairs` (log-mu-ref (theta_idx, eta_idx) pairs)
  - `OMEGA_DIAG_FLOOR` (1e-6, with the MH proposal-collapse rationale)
Both `saem.rs` and `impmap.rs` now call the single copy; the dedicated unit
tests move to `em_common` (their canonical home) and cover both estimators.

Efficiency (IMPMAP hot loop):
  - `subject_is_draws` retained samples now live in one flat `Vec<f64>` of
    length K·d per subject instead of K small `vec![0.0; d]` allocations per
    subject per iteration (~7.5M short-lived allocs at K=500, N=100, 150 iters).
  - `params_k` is built once and the three changing fields (theta, sigma.values,
    omega) are refreshed in place each iteration instead of rebuilding the full
    16-field struct; the final-params bundle reuses it too.
  - Drop the redundant `eta_hats.clone()` + second `last_eta_hats` binding; a
    single `prev_etas` carries the modes forward and out to the final EBE pass.

Simplification: drop the redundant `value.trim()` in the `impmap_proposal_df`
parser branch (`value` is trimmed at the top of `apply_fit_option`).

Deferred (noted in #273, larger blast radius): a shared packed-space M-step
helper, and threading the mode's ipreds/r_diag out of `run_inner_loop_warm` to
avoid `compute_posterior_hessian`'s redundant re-prediction.

Numerics unchanged: the IMPMAP↔NONMEM warfarin match test and the
IMPMAP→FOCEI convergence test pass under `--features slow-tests`.
