+++ src/estimation/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod em_common;
 pub mod gauss_newton;
 pub(crate) mod hmc;
 pub mod impmap;
